[xwang32 & ymeng3] Fix index out of range exception
git-svn-id: https://bgc-mengyu.dir.slb.com:441/svn/CodeWorldCup/trunk@33 c40e7b8e-7e2c-fe40-b653-1a1a820b3dea

+++ worldcup/domain_s2/strategy.py
@@ -23,7 +23,7 @@ class RandomExploration:
                 return Occupied(self.block_map, last_action.x, last_action.y)
         elif last_action.type == ActionType.EXPLORE:
             if is_last_action_success:
-                return Explored(self.block_map, last_action.x, last_action.y, int(last_operation_value))
+                return Explored(self.block_map, last_action.x, last_action.y, float(last_operation_value))
             else:
                 # TODO: insufficient funds
                 pass
